requirements.txt
aiohttp==3.6.2
argcomplete==1.12.0
async-timeout==3.0.1
attrs==19.3.0
boto3==1.14.34
botocore==1.17.34
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
docutils==0.15.2
durationpy==0.5
Flask==1.1.2
future==0.18.2
hjson==3.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
multidict==4.7.6
pip-tools==5.3.1
placebo==0.9.0
pyee==7.0.2
python-dateutil==2.6.1
python-slugify==4.0.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
slack==0.0.2
slackclient==2.7.3
slackeventsapi==2.2.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.48.2
troposphere==2.6.2
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
yarl==1.5.1
zappa==0.51.0
